Separate the position and shape of models
Previously, the graphical representation of an object was fully
determined by the WorldObjectDescriptor. This refactor wraps this into
an ModelDescriptor object which contains only the varying
position/orientation used for creating modelview matrices.  This will
enable a number of performance optimizations, since now
WorldObjectDescriptors do not have to be recreated every time the
position/orientation of an object changes.