Add dedicated class to store position of models
This performance optimisation prevents us from having to recreate every
modelview matrix on every timestep. The position part of ModelDescriptor
is now encapsulated in a dedicated PositionDescriptor class. This allows
the PositionDescriptor to be cached by objects that don't change
position.  We can now also attach a modelview matrix to the
PositionDescriptor, and retrive this again on the next frame (instead of
recreating the matrix) if we encounter the same PositionDescriptor on a
subsequent frame.