Combine submodels with of same material
Add an optimized method to CompositeModelBuilder that first flattens the
model hierachy (collecting all the sub-models from children which are
also CompositeModelBuilder) and then combines the vertex data of all
PrimitiveModelBuilders of the same material. This significantly reduces
the size of the model tree and consequently the number of draw calls and
other operations.