Restore determinism
Some game mechanics had been added which rely on the ordering in which
drones are updated (e.g. two drones try to harvest the same mineral,
drone is concurrently harvesting minerals and given minerals). To
eliminate conflicts like this an ordering is now imposed on all drones
by means of a randomly generated number, with the id used as tiebraker.
(Using just the id would not be fair, since then one of the players
drones would always be updated first.)