Improve VisionTracker performance
As before, VisionTracker uses SquareGrid to partition 2D space into
square shaped sectors. This allows efficient computation (and update) of
a list of nearby objects for each object. In addtion, each object keeps
a priority queue of nearby objects, sorted by the earliest possible time
that each object may leave/enter the sight radius. This time is computed
from the distance of the objects, and an upper bound of the speed of the
objects.

Compared to the simple previous implementation, this results in more
than 10x speedup for VisionTracker. Overall run time is reduced by ~15%.