Fix a bug causing infinite loop in PhysicsEngine
The code for handling wall collisions was biased towards flipping the y
component of the velocity. This caused an infinite loop when a drone
collided with both walls simultaneously (only the y component was
flipped back and forth, and never the x component, meaning the drone was
always still on a collision course).